requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

modelscope==1.19.0
numpy==1.24.3
openpyxl==3.1.5
python-pptx==1.0.2
torch==2.4.1
